changed type from float to double

+++ app/src/main/java/se/miun/dt170g/laboration01/WeatherValues.java
@@ -2,16 +2,16 @@ package se.miun.dt170g.laboration01;
 
 public class WeatherValues {
 
-    float temperature_degree;
+    double temperature_degree;
     String temperature_unit = "Celsius";
 
-    float wind_speed;
+    double wind_speed;
     String wind_unit = "mps";
     String wind_direction;
 
-    float cloudiness_percentage;
+    double cloudiness_percentage;
 
     String percipitation_unit = "mm";
-    float percipitation_min;
-    float percipitation_max;
+    double percipitation_min;
+    double percipitation_max;
 }
